+++ src/vm/tests/mock_tests/mod.rs
@@ -24,7 +24,8 @@ pub(crate) mod mock_test_prelude {
 }
 
 mod mock_test_allocate_align_offset;
-mod mock_test_allocate_no_gc_oom_on_acquire;
+mod mock_test_allocate_no_gc_oom_on_acquire_allow_oom_call;
+mod mock_test_allocate_no_gc_oom_on_acquire_no_oom_call;
 mod mock_test_allocate_no_gc_simple;
 mod mock_test_allocate_nonmoving;
 mod mock_test_allocate_overcommit;
